requirements.txt
# external requirements (pinned for reproducibility)
importlib-metadata==4.13.0
importlib-resources==5.12.0
flake8==6.1.0
python-dotenv==1.0.1
pytest==7.4.4
numpy==1.24.4
pandas==1.5.3
seaborn==0.12.2
matplotlib==3.7.3
scipy==1.10.1
torch==1.13.1
pytorch-lightning==1.4.5
torchmetrics==0.6.0
hydra-core==1.3.2
hydra_colorlog==1.2.0
mlflow==2.12.2
scikit-learn==1.2.2
tables==3.9.2
ray[tune]==2.9.3
mne==1.4.2
wfdb==4.1.2
pyedflib==0.1.37

git+https://github.com/fadel/pytorch_ema
